feat: seperated update and delete feature
