+++ src/js/editPage.js
@@ -12,12 +12,12 @@ class EditBook {
             <input type="text" name="title" placeholder="Title" class="book-input title" required />
             <input type="text" name="author" placeholder="Author" class="book-input author" required />
             <select name="genre" class = "book-input genre">
-                  <option value = "choose" selected>Select Genre</option>
+                  <option value = "nonSelected" selected>Select Genre</option>
                   <option value = "Fiction">Fiction</option>
-                  <option value = "Non-Fiction">Non-Fiction</option>
+                  <option value = "nonFiction">Non-Fiction</option>
             </select>
             <select name="category" class = "book-input category">
-              <option value = "Non Selected" selected>Select Category</option>                            
+              <option value = "nonSelected" selected>Select Category</option>                            
             </select>
             <button type="submit" class="form-button">Update</button>
           </form>
