+++ src/js/utils/formCategorySelector.js
@@ -1,5 +1,5 @@
 const fictionCategory = `
-    <option value = "Non Selected" selected>Select Category</option>
+    <option value = "nonSelected" selected>Select Category</option>
     <option value = "Action & Adventure">Action & Adventure</option>
     <option value = "Comic">Comic</option>
     <option value = "Classic">Classic</option>
@@ -12,7 +12,7 @@ const fictionCategory = `
 `;
 
 const nonFictionCategory = `
-    <option value = "Non Selected" selected>Select Category</option>              
+    <option value = "nonSelected" selected>Select Category</option>              
     <option value = "Biographies">Biographies</option>
     <option value = "Business">Business</option>
     <option value = "Computer & Tech">Computer & Tech</option>
@@ -25,19 +25,19 @@ const nonFictionCategory = `
 `;
 
 const emptyCategory = `
-    <option value = "Non Selected" selected>Select Category</option>
+    <option value = "nonSelected" selected>Select Category</option>
 `;
 
 const selectCategory = () => {
   const genreChange = document.querySelector('.genre');
   const category = document.querySelector('.category');
-  if (genreChange.value === 'Fiction') {
-    category.innerHTML = fictionCategory;
-  } else if (genreChange.value === 'Non-Fiction') {
-    category.innerHTML = nonFictionCategory;
-  } else if (genreChange.value === 'Non Selected') {
-    category.innerHTML = emptyCategory;
-  }
+  const genreValue = {
+    Fiction: fictionCategory,
+    nonFiction: nonFictionCategory,
+    nonSelected: emptyCategory,
+  };
+  const newChild = genreValue[genreChange.value];
+  category.innerHTML = newChild;
 };
 
 export default selectCategory;
